Refactor GameObject to SerializedObject
This helps avoid data duplication.

Now SerializedObjects are to be used solely to save/load or as a form of
entity copying for things like moving tiles.

The data is somewhat duplicated through the SerializedObjectType,
but this information should be used as reference only to identify the
type of the entity, and not to use its inner data. The data of any
entity should only ever be fetched though their corresponding
components.